Improvements to config files.
Configs can now be merged. This allows for easy overriding of default
settings. An example may be that you have some default settings, and
then users can set their own settings by modifying only the things that
they want to change and leave everything else alone.

Another use case might be mods, where each mod might want to have it's
own isolated set of settings which leave the base uneffected. When the
mod is loaded, the mod's config is also loaded and the settings are
applied. When the mod is not loaded, that config is never loaded and
therefore all the base settings remain.