Start work on wiring up some basic config settings.
With this commit, e_client_config_init() has changed so that it takes
only a pointer to the owning `e_engine` object and a string containing
the name of the section to load configuration settings for the client.

At the moment the only config settings that are loaded from files is
window title and resolution.

As an example, if you had one client that represented your game, you
can have a section called "game":

    game =
    {
        title = "My Game";
    }

And then you could have another client that represents the editing
tools which might have the section called "editor":

    editor =
    {
        title = "My Game Editor";
    }

If a properly is explicitly set via the `e_client_config` object,
they will take precedence.