Add low-level wrapper for sockets.
The Windows build requires an explicit call to e_winsock_init() which
is where the ws2_32.dll is loaded and WinSock initialized. Likewise, a
corresponding call to e_winsock_uninit() is also required.

e_winsock_init() is where the global functions are bound. It *must* be
called before any socket functions are called.